Mobile_Automation_Python/Elli/requirements.txt
Appium-Python-Client
iniconfig==2.0.0
pip
pycparser==2.22
pytest==8.2.2
selenium==4.21.0
